tests/wpt/tests/fenced-frame/report-event.https.html

<!DOCTYPE html>
<title>Test that window.fence.reportEvent() succeeds in a fenced frame.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="resources/utils.js"></script>

<body>
  <script>
  promise_test(async(t) => {
    const fencedframe = await attachFencedFrameContext({generator_api: 'fledge',
                                                        register_beacon: true});

    // This page will call reportEvent twice: Once for an enum event, and once
    // for a custom URL event. Both beacons are expected to send successfully.
    await fencedframe.execute(() => {
      const destination_enum_event = {
        eventType: 'click',
        eventData: 'enum',
        destination: ['buyer']
      }
      window.fence.reportEvent(destination_enum_event);

      const destination_url = new URL(BEACON_URL + "?type=url",
          get_host_info().HTTPS_ORIGIN);

      const destination_url_event = {
        destinationURL: destination_url
      }
      window.fence.reportEvent(destination_url_event);
    });

  let enum_data = await nextBeacon('click', 'enum');
  assert_equals(enum_data, location.origin);

  let url_data = await nextBeacon('url', '<No data>');
  assert_equals(url_data, '<No data>');

  }, 'Test that window.fence.reportEvent() succeeds in a fenced frame.');
  </script>
  </body>
